requirements.txt
numpy
pillow
requests
pika
